refactor: pushNotification 훅 내부 함수들 관심사에 따른 분리
