+++ src/fsd_entities/notification/api/post.ts
@@ -26,7 +26,6 @@ export const updateFCMToken = async (token: string): Promise<void> => {
   try {
     await API.post(URI);
   } catch (error) {
-    toast.error('FCM 토큰 업데이트 실패: 서버 응답 오류');
     throw error;
   }
 };
