+++ src/fsd_entities/notification/model/usePushNotification.ts
@@ -1,83 +1,29 @@
-import { getMessaging, getToken } from 'firebase/messaging';
-
 import { getFCMToken } from '../api/get';
 import { updateFCMToken } from '../api/post';
 import { toast } from 'react-hot-toast';
+import { STORAGE_KEYS } from '@/fsd_shared/configs';
+import { detectDeviceType, getClientFCMToken } from '@/fsd_shared/model';
+import { requestNotificationPermission } from '@/fsd_shared/utils';
 
-const FCM_CHECKED_KEY = 'fcm_checked';
-
-export type DeviceType = 'android' | 'ios' | 'ipad' | 'desktop';
+const FCM_CHECKED_KEY = STORAGE_KEYS.FCM_CHECKED;
 
 export const usePushNotification = () => {
-  const getDeviceType = (): DeviceType => {
-    const ua = navigator.userAgent.toLowerCase();
-
-    // Android
-    if (/android/.test(ua)) return 'android';
-
-    // iPhone, iPod
-    if (/iphone|ipod/.test(ua)) return 'ios';
-
-    // iPad (modern iPadOS가 userAgent에 Mac처럼 보이는 경우도 커버)
-    const isIpad = /ipad/.test(ua) || (navigator.platform === 'MacIntel' && navigator.maxTouchPoints > 1);
-    if (isIpad) return 'ipad';
-
-    return 'desktop';
-  };
-
-  // 알림 권한 요청
-  const requestPushPermission = async (): Promise<NotificationPermission> => {
-    if (Notification.permission === 'granted') {
-      return 'granted';
-    } else if (Notification.permission === 'denied') {
-      return 'denied';
-    } else {
-      return await Notification.requestPermission();
-    }
-  };
-
-  // FCM 토큰 가져오기
-  const getClientFCMToken = async (): Promise<string | null> => {
-    try {
-      // 권한 요청
-      const permission = await requestPushPermission();
-      if (permission !== 'granted') {
-        return null;
-      }
-
-      const messaging = getMessaging();
-      const currentToken = await getToken(messaging, {
-        vapidKey: process.env.NEXT_PUBLIC_FB_VAPID_KEY,
-      });
-      if (!currentToken) {
-        return null;
-      }
-      // 토큰이 있는 경우 저장
-      return currentToken;
-    } catch {
-      throw new Error('Failed to get FCM token: Unknown error');
-    }
-  };
 
   // 토큰 동기화
   const compareFCMToken = async (): Promise<void> => {
     try {
-      const deviceType = getDeviceType();
+      const deviceType = detectDeviceType();
       if (deviceType === 'desktop') {
         return;
       }
-
-      // 이미 체크된 경우 스킵
-      const checked = localStorage.getItem(FCM_CHECKED_KEY);
-      if (checked) {
-        return;
-      }
-      localStorage.setItem(FCM_CHECKED_KEY, 'true');
       
-      // 알림 권한 요청
-      await Notification.requestPermission();
-
-      if (Notification.permission !== 'granted') {
+      // 알림 권한 granted 아닌
+      if (Notification.permission === 'default') {
+        const permission = await requestNotificationPermission();
+        if (permission !== 'granted') {
+          return;
+        }
+      } else if (Notification.permission === 'denied') {
         return;
       }
 
@@ -86,6 +32,7 @@ export const usePushNotification = () => {
 
       // 로컬에서 토큰 못 가져오는 경우 early return
       if (!clientFCMToken) {
+        toast.error('알림 설정 실패');
         return;
       }
 
@@ -96,9 +43,9 @@ export const usePushNotification = () => {
       if (!fcmToken.includes(clientFCMToken)) {
         await updateFCMToken(clientFCMToken);
         toast.success('알림 설정을 허용하였습니다.');
-      }
+      } 
     } catch (error) {
-      toast.error('FCM 토큰 동기화 실패: 서버 응답 오류');
+      return ;
     }
   };
 
@@ -108,8 +55,7 @@ export const usePushNotification = () => {
   };
 
   return {
-    getDeviceType,
-    requestPushPermission,
+    requestPushPermission: requestNotificationPermission,
     compareFCMToken,
     resetFCMToken,
   };
