+++ src/fsd_shared/configs/constants/firebase.ts
@@ -1,0 +1,4 @@
+export const FIREBASE_CONFIG = {
+    vapidKey: process.env.NEXT_PUBLIC_FB_VAPID_KEY,
+  } as const;
+
