+++ src/fsd_shared/configs/constants/index.ts
@@ -5,3 +5,5 @@ export * from './notificationTabs';
 export * from './locker';
 export * from './date';
 export * from './regex';
+export * from './firebase';
+export * from './storage';
