+++ src/fsd_shared/configs/constants/storage.ts
@@ -1,0 +1,4 @@
+export const STORAGE_KEYS = {
+    FCM_CHECKED: 'fcm_checked',
+  } as const;
+
