+++ src/fsd_shared/configs/index.ts
@@ -5,3 +5,4 @@ export * from './query-key';
 export * from './constants';
 export * from './api';
 export * from './permission';
+export * from './constants';
