+++ src/fsd_shared/model/index.ts
@@ -1,2 +1,4 @@
 export * from './format';
 export { sanitizeHtml } from './sanitizeHtml';
+export { detectDeviceType } from './device';
+export { getClientFCMToken } from './fcmToken';
