+++ src/fsd_shared/utils/index.ts
@@ -1,1 +1,2 @@
 export * from './tokenManager';
+export * from './notificationPermission';
